+++ utils/image_reader_pgn.py
@@ -16,8 +16,8 @@ def image_scaling(img, label, edge):
     """
     
     scale = tf.random_uniform([1], minval=0.5, maxval=2.0, dtype=tf.float32, seed=None)
-    h_new = tf.to_int32(tf.multiply(tf.to_float(tf.shape(img)[0]), scale))
-    w_new = tf.to_int32(tf.multiply(tf.to_float(tf.shape(img)[1]), scale))
+    h_new = tf.to_int32(tf.multiply(tf.cast(tf.shape(img)[0], tf.float32), scale))
+    w_new = tf.to_int32(tf.multiply(tf.cast(tf.shape(img)[1], tf.float32), scale))
     new_shape = tf.squeeze(tf.stack([h_new, w_new]), squeeze_dims=[1])
     img = tf.image.resize_images(img, new_shape)
     label = tf.image.resize_nearest_neighbor(tf.expand_dims(label, 0), new_shape)
@@ -53,8 +53,8 @@ def image_mirroring(img, label, label_rev, edge):
 def random_resize_img_labels(image, label, resized_h, resized_w):
 
     scale = tf.random_uniform([1], minval=0.75, maxval=1.25, dtype=tf.float32, seed=None)
-    h_new = tf.to_int32(tf.multiply(tf.to_float(resized_h), scale))
-    w_new = tf.to_int32(tf.multiply(tf.to_float(resized_w), scale))
+    h_new = tf.to_int32(tf.multiply(tf.cast(resized_h, tf.float32), scale))
+    w_new = tf.to_int32(tf.multiply(tf.cast(resized_w, tf.float32), scale))
 
     new_shape = tf.squeeze(tf.stack([h_new, w_new]), squeeze_dims=[1])
     img = tf.image.resize_images(image, new_shape)
